% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEnvironmental.R
\name{addEnvironmental}
\alias{addEnvironmental}
\title{Add environmental data to core westbrook data}
\usage{
addEnvironmental(coreData, tempData)
}
\arguments{
\item{coreData}{a data.frame created using createCoreData}
}
\value{
Environmental data added to a coreData
}
\description{
Add environmental data to core westbrook data
}
\details{
\strong{Options of columns to add}
}

