% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addKnownZ.R
\name{addKnownZ}
\alias{addKnownZ}
\title{Adds known state to data.frame created with createCmrData()}
\usage{
addKnownZ(cmrData, knownDead = T, useAntenna = F)
}
\arguments{
\item{cmrData}{A data.frame created using createCmrData()}

\item{knownDead}{Logical indicating whether to set known state based on tags known to be dead}
}
\value{
A data.frame including \code{$knownZ} which is the known state (0=not born,1=alive,2=dead,NA=unknown)
}
\description{
Adds known state to data.frame created with createCmrData()
}
