% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSampleProperties.R
\name{addSampleProperties}
\alias{addSampleProperties}
\title{Add immutable sample properties to a data frame}
\usage{
addSampleProperties(data, defaultColumns = T, columnsToAdd = NULL)
}
\arguments{
\item{data}{A data.frame created with createCoreData(), which must include sampleNumber as a column}

\item{columnsToAdd}{A character vector of columns to inlcude; can replace or add to baseColumns}

\item{basecolumns}{Logical: include default columns? year,season,medianDate}
}
\value{
A data.frame appended with sample properties
}
\description{
Add immutable sample properties to a data frame
}

