% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSectionProperties.R
\name{addSectionProperties}
\alias{addSectionProperties}
\title{Add section properties (stream width by section for each sample) to a data frame. Missing data are estimated based on log(width) ~ log(stream flow))}
\usage{
addSectionProperties(
  data,
  defaultColumns = T,
  columnsToAdd = NULL,
  estimateWidthNA = T
)
}
\arguments{
\item{data}{A data.frame created with createCoreData(), which must include river and section as a column}

\item{columnsToAdd}{A character vector of columns to inlcude; can replace or add to baseColumns}

\item{basecolumns}{Logical: include default columns? river,sampleName,width, maximum_depth}

\item{estimatedWidth}{Logical: estimate widths for missing data? Default = T}
}
\value{
A data.frame appended with sample properties
}
\description{
Add section properties (stream width by section for each sample) to a data frame. Missing data are estimated based on log(width) ~ log(stream flow))
}
