% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTagProperties.R
\name{addTagProperties}
\alias{addTagProperties}
\title{Download data from West Brook database}
\usage{
addTagProperties(coreData, columnsToAdd = c("date_known_dead",
  "lastAntennaDetection", "species", "firstCaptureSample", "lastCaptureSample",
  "cohort"))
}
\arguments{
\item{coreData}{a data.frame created using createCoreData}

\item{columnsToAdd}{A character vector of columns to inlcude; can replace or add to baseColumns}
}
\value{
Immutable tag properties added to a data.frame
}
\description{
Download data from West Brook database
}
\details{
\strong{Options of columns to add}
date_known_dead, lastAntennaDetection, species, firstCaptureSample, lastCaptureSample, cohort
}

