% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCmrData.R
\name{createCmrData}
\alias{createCmrData}
\title{Create data strucuture for capture-mark-recapture analyses
Creates encounter histories for each individual and indices for reference within jags}
\usage{
createCmrData(coreData, dateStart = as.POSIXct("1900-01-01"),
  dateEnd = as.POSIXct("2100-01-01"), maxAgeInSamples = 20,
  censorDead = F, censorEmigrated = T, modelType = "CJS", inside = T)
}
\arguments{
\item{coreData}{A data.frame created with \code{createCoreData()}}

\item{maxAgeInSamples}{An age in samples (age/4) after which encounter histories are censored}

\item{startDate}{A Date or POSIXct time to pad samples back to}

\item{endDate}{A A Date or POSIXct time to pad samples forward to}
}
\value{
A data.frame containing encounter histories (\code{$enc}) with all samples 
 within the range defined. \code{$sampleIndex} and \code{$tagIndex} are also created
 for refence within jags objects.
}
\description{
Create data strucuture for capture-mark-recapture analyses
Creates encounter histories for each individual and indices for reference within jags
}

