% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCoreData.R
\name{createCoreData}
\alias{createCoreData}
\title{Download data from West Brook database}
\usage{
createCoreData(sampleType = NULL, baseColumns = T, columnsToAdd = NULL,
  includeUntagged = F)
}
\arguments{
\item{sampleType}{A character vector of sampling types to be included in output, options are: electrofishing,stationaryAntenna,portableAntenna,trap,seine,snorkel}

\item{baseColumns}{Logical: include default columns? tag,detection_date}

\item{columnsToAdd}{A character vector of columns to inlcude; can replace or add to baseColumns}

\item{includeUntagged}{Logical: include untagged fish?}
}
\value{
a data.frame of fish data
}
\description{
Download data from West Brook database
}
\details{
\strong{Column names for tables}
These are the options for columns to add depending on the sampleType chosen. 
 When multiple sampleTypes are selected, columns that only exist within one are given NA for the other sampleType(s)

\emph{captures:}

tag, fishNumber, species, cohort, sampleNumber, detectionDate, seasonNumber, river, area, section, observedLength, observedWeight, survey, sampleName

\emph{stationaryAntenna:}

tag, detectionDate, river, area, section, survey, sampleName, readerId, sampleType, aliveOrDead, instance, arrival, departure, comment

\emph{portableAntenna:}

tag, detectionDate, river, area, section, survey, sampleName, readerId, sampleType, aliveOrDead, instance, pass, quarter, leftOrRight, habitat, cover, justification, comment
}

