% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJAGSData.R
\name{createJAGSData}
\alias{createJAGSData}
\title{Create data needed for JAGS runs}
\usage{
createJAGSData(coreData, modelType = "CJS")
}
\arguments{
\item{coreData}{a data.frame created using createCoreData}
}
\value{
List of evalRows and firstObsRows
}
\description{
Create data needed for JAGS runs
}
\details{
\strong{Options of columns to add}
}

