% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillSizeLocation.R
\name{fillSizeLocation}
\alias{fillSizeLocation}
\title{Fills in data for location and length when a fish was not observed using some simple assumptions
For location, a fish is assumed to remain in the same location until it is observed elsewhere. For length, a linear interpolation is used to fill lengths between observations and the average by ageInSamples/river is used after the last observation.}
\usage{
fillSizeLocation(data, size = T, location = T)
}
\arguments{
\item{data}{A data.frame created with createCoreData() and createCmrData()}
}
\value{
Data for location and size even when fish are not observed.
}
\description{
Fills in data for location and length when a fish was not observed using some simple assumptions
For location, a fish is assumed to remain in the same location until it is observed elsewhere. For length, a linear interpolation is used to fill lengths between observations and the average by ageInSamples/river is used after the last observation.
}

