% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCase.R
\name{fillUnderscore}
\alias{fillUnderscore}
\title{Convert camel-case string to underscore separated}
\usage{
fillUnderscore(x)
}
\arguments{
\item{x}{Character (vector) to be converted to camel-case}
}
\value{
A string (vector) with underscore separated words
}
\description{
Convert camel-case string to underscore separated
}

