% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFish.R
\name{getFish}
\alias{getFish}
\title{Test function for extracting data}
\usage{
getFish(species = c("bkt", "bnt", "ats"), tagged = T, sampleType = NA,
  baseColumns = T, columnsToAdd = NULL)
}
\description{
Test function for extracting data
}

