% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbConnector.R
\name{reconnect}
\alias{reconnect}
\title{Reconnect to the West Brook database}
\usage{
reconnect()
}
\value{
A connection to the West Brook database
}
\description{
Tests the connection to the database and runs wbConnect() if the connection has not been established or has expired
}

