% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEnvironmental.R
\name{addEnvironmental}
\alias{addEnvironmental}
\title{Add environmental data to core westbrook data}
\usage{
addEnvironmental(coreData, sampleFlow = F, funName = "mean")
}
\arguments{
\item{coreData}{a data.frame created using createCoreData}

\item{sampleFlow}{Logical indicating whether to add the discharge on detectionDate to the output}

\item{fun}{A quoted name of a function used to characterize the environmental data between samples (e.g., mean, median, min, max). Custom functions can be used but must have two arguments: x and na.rm.}
}
\value{
Environmental data added to a coreData
}
\description{
Add environmental data to core westbrook data
}
\details{
View this function as a template to be altered to get different env variables or different summaries. / This function add columns of mean daily temperature and mean daily flow to coreData. For each individual, it returns the mean env values between capture dates or between median capture dates of all individuals of a fish was not captured on an occasion.
The internal function getIntervalMean returns means for the observed river or for all rivers if an individual was unobserved.
Could alter getIntervalMean to return data for the mean daily min or max env data or for data from the last observed river, instead of all rivers.
}

