% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSampleProperties.R
\name{addSampleProperties}
\alias{addSampleProperties}
\title{needs to include year,season,medianDate,proportionSampled}
\usage{
addSampleProperties(data)
}
\description{
medianDate should fill in observedDate, add option to addJulian version,
year should be minimum year for the sample
}

