% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTagProperties.R
\name{addTagProperties}
\alias{addTagProperties}
\title{Add immutable tag properties to core westbrook data}
\usage{
addTagProperties(coreData, columnsToAdd = c("species", "cohort"))
}
\arguments{
\item{coreData}{a data.frame created using createCoreData}

\item{columnsToAdd}{A character vector of columns to inlcude; can replace or add to baseColumns}
}
\value{
Immutable tag properties added to a data.frame
}
\description{
Add immutable tag properties to core westbrook data
}
\details{
\strong{Options of columns to add}
dateKnownDead, lastAntennaDetection, species, firstCaptureSample, lastCaptureSample, cohort, sex, familyId
}

