% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCoreData.R
\name{createCoreData}
\alias{createCoreData}
\title{Download data from West Brook database}
\usage{
createCoreData(sampleType = "electrofishing", species = c("bkt", "bnt",
  "ats"), baseColumns = T, columnsToAdd = NULL, includeUntagged = F)
}
\arguments{
\item{sampleType}{A character vector of sampling types to be included in output, options are: electrofishing,stationaryAntenna,portableAntenna,trap}

\item{species}{A character vector of species to include, possible codes: bkt,bnt,ats,bnd}

\item{baseColumns}{Logical: include default columns? tag,detection_date}

\item{columnsToAdd}{A character vector of columns to inlcude; can replace or add to baseColumns}

\item{includeUntagged}{Logical: include untagged fish?}
}
\description{
Download data from West Brook database
}

